fix: Agregar control Corrección tonta para evitar NullExceptions Issue: #1

+++ fuentes/eeutil-pdfatools/src/main/java/es/gob/ad/eeutil/eeutilpdfatools/services/PdfaToolsService.java
@@ -264,7 +264,7 @@ public class PdfaToolsService {
     } catch (Exception e) {
       throw new EeutilPdfaToolsException(e.getMessage(), e);
     } finally {
-      if (process.isAlive()) {
+      if (process != null && process.isAlive()) {
         // si despues de la excepcion sigue vivo lo forzamos a cerrar
         process.destroyForcibly();
       }
